Created connection file to database. Basic Authorization logic with returning of jwt token.
Configuration file with function for creating session with database was created. Also there is test to check connection to database. In Server.rs was added function login with same endpoint, which looks for user in database, checks on password, and in case of success returns json web token. currently jwt_functions have function to encode and decode tokens by secret and algorithm hs512. In models is saved structure which send client to server for request authentication and user structure to work with data from database on server side. All data to connect to database or secret to work with jwt tokens is stored in .env file, which will be in gitignore list in safe intentions.